Rename AttendanceModal -> AttendanceModalContent
