+++ app/components/UserAttendance/AttendanceModalContent.tsx
@@ -8,7 +8,7 @@ import { ProfilePicture } from 'app/components/Image';
 import Flex from 'app/components/Layout/Flex';
 import type { ID } from 'app/store/models';
 import type { PublicUser } from 'app/store/models/User';
-import styles from './AttendanceModal.css';
+import styles from './AttendanceModalContent.css';
 
 export type Registration = {
   id: ID;
@@ -48,7 +48,7 @@ const Tab = ({ name, index, activePoolIndex, togglePool }: TabProps) => (
   </Button>
 );
 
-const AttendanceModal = ({
+const AttendanceModalContent = ({
   title = 'Status',
   pools,
   togglePool,
@@ -141,4 +141,4 @@ const AttendanceModal = ({
   );
 };
 
-export default AttendanceModal;
+export default AttendanceModalContent;
