+++ app/components/UserAttendance/index.ts
@@ -1,4 +1,3 @@
 export { default as AttendanceStatus } from './AttendanceStatus';
-export { default as AttendanceModal } from './AttendanceModal';
 export { ModalParentComponent } from './withModal';
 export { default as withModal } from './withModal';
